% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutes.R
\name{get_mutes}
\alias{get_mutes}
\title{Get list of mutes for the authenticated user}
\usage{
get_mutes(token = NULL, n = 5000)
}
\arguments{
\item{token}{Token}

\item{n}{Maximum number of user_ids to retrieve}
}
\value{

}
\description{
Get list of mutes for the authenticated user
}
